Initial commit of first part of code
LICENSE is MIT from github,

ai/game.php is the main RESTful API endpoint that handles the game.
Currently everything (making players, joining games etc.) is in this one
file, which should work fine. It will probably be good to split into
separate files later. Also, the documentation for calls is all hidden in
the game comments at the moment.

ai_game_db.sql is a blank copy of the database for installation. It does
not include roles, users etc as defined in the beginning of game.php, so
that will need to get configured.

analysis/win_loss_tie.c is a short program to calculate the possible
configurations of games of a certain length. For example, with 7 hands,
the percentages are 41, 41, 18 for win, lose, tie. Also outputs data for
all configurations in csv.

setup.sh copies ai/ recursively into /var/www/html/ for installation on
linux.