Random Play Daemon
To be executed when a player wants to play against the random daemon.

Passed the game_name via the command line.

Waits 1 second, then 2 seconds, then 4 seconds etc. between turns if the
other player hasn't made a play yet. If the daemon waits 256 seconds and
the other player hasn't played, it gets bored and quits.